+++ client/plugins/spotify/endpoints/search/index.ts
@@ -1,0 +1,7 @@
+import { Context } from '@nuxt/types';
+
+import { searchItems } from './searchItems';
+
+export const search = (context: Context) => ({
+  searchItems: searchItems(context),
+});
